created snackbar at app level
