+++ app/src/App.js
@@ -1,25 +1,30 @@
 import React from 'react'
 import { Routes, Route } from 'react-router-dom'
+import Aboutus from './Pages/AboutUs'
 import MenuAppBar from './Components/MenuAppBar'
 import { ThemeProvider } from '@mui/material'
 import theme from './Components/Theme'
 import Home from './Pages/Home'
 import Team from './Pages/Team'
-import Aboutus from './Pages/AboutUs'
 import Menufooter from './Components/MenuFooter'
 import PrivacyPolicy from './Pages/PrivacyPolicy'
+import SnackbarStateProvider from './Providers/SnackbarStateProvider'
+import Snackbar from './Components/MuiComponents/Snackbar'
 
 const App = () => {
   return (
     <>
       <ThemeProvider theme={theme}>
         <MenuAppBar />
-        <Routes>
-          <Route path="/" element={<Home />} />
-          <Route path="/team" element={<Team />} />
-          <Route path="/about-us" element={<Aboutus />} />
-          <Route path="/privacy-policy" element={<PrivacyPolicy />} />
-        </Routes>
+        <SnackbarStateProvider>
+          <Routes>
+            <Route path="/" element={<Home />} />
+            <Route path="/team" element={<Team />} />
+            <Route path="/about-us" element={<Aboutus />} />
+            <Route path="/privacy-policy" element={<PrivacyPolicy />} />
+          </Routes>
+          <Snackbar />
+        </SnackbarStateProvider>
         <Menufooter />
       </ThemeProvider>
     </>
