+++ app/src/Providers/SnackbarStateProvider.js
@@ -1,0 +1,14 @@
+import React from 'react'
+export const SnackbarContext = React.createContext({})
+
+const SnackbarStateProvider = ({ children }) => {
+    const [openToast, setOpenToast] = React.useState(false)
+
+    return (
+        <SnackbarContext.Provider value={{ openToast, setOpenToast }}>
+            {children}
+        </SnackbarContext.Provider>
+    )
+}
+
+export default SnackbarStateProvider
